LandslidePredictionSystem/requirements.txt
flask
requests
ibm_watson_machine_learning
python-decouple
